separate data utils from other utils
